requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Django==5.2.3
fastapi==0.115.14
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
motor==3.3.2
uvicorn==0.27.1
python-dotenv==1.0.0
requests==2.31.0
